refactor(sa-base): complete core module refactoring with new foundation architecture
Complete Phases 1-6 of core module refactoring plan, transforming the
monolithic foundation/core module into 5 focused modules with clear
separation of concerns. This establishes a cleaner architecture while
maintaining backward compatibility through deprecated bridge classes.

## New Foundation Modules Created

1. **foundation/domain** - Core domain objects
   - ResponseDTO, PageResult, PageParam, RequestUser
   - ErrorCode hierarchy (ErrorCode, UserErrorCode, SystemErrorCode, UnexpectedErrorCode)
   - BaseEnum interface and DataTypeEnum
   - Constants (StringConst, RequestHeaderConst)
   - Package: net.lab1024.sa.foundation.domain.*

2. **foundation/validation** - Validation framework
   - @CheckEnum annotation and EnumValidator
   - SmartEnumUtil for enum utilities
   - ValidateList and ValidateData
   - Package: net.lab1024.sa.foundation.validation.*

3. **foundation/json** - JSON serialization
   - LongJsonSerializer, LongJsonDeserializer
   - EnumSerializer and EnumSerialize
   - BigDecimalNullZeroSerializer
   - JsonUtil for JSON operations
   - Package: net.lab1024.sa.foundation.json.*

4. **foundation/ip-geolocation** - IP geolocation utilities
   - IpGeolocationUtil (renamed from SmartIpUtil)
   - Separated from core to isolate ip2region dependency
   - Package: net.lab1024.sa.foundation.ipgeo.*

## Infrastructure Module Enhancements

- **infrastructure/web**: Added SmartRequestUtil, SmartResponseUtil
- **infrastructure/mybatis**: Added SmartPageUtil for pagination
- **infrastructure/swagger**: Added SwaggerTagConst for tag management

## Key Changes

### Module Dependencies (28+ modules updated)
All modules now reference specific foundation modules instead of the
monolithic foundation/core:

```kotlin
// Before
dependencies {
    api(project(":sa-base:foundation:core"))
}

// After
dependencies {
    api(project(":sa-base:foundation:domain"))
    api(project(":sa-base:foundation:validation"))
    api(project(":sa-base:foundation:json"))
    // Optional: ip-geolocation only when needed
}
```

Updated modules:
- Infrastructure: datasource, mybatis, web, redis, swagger, token, devtools
- Foundation: api-encrypt, repeat-submit, cache, mq
- Support: 17+ modules (changelog, codegenerator, config, datatracer, dict,
  feedback, file, heartbeat, helpdoc, job, loginlog, mail, message,
  operatelog, reload, serialnumber, table, etc.)
- sa-admin: Main application module

### Backward Compatibility Strategy
Created deprecated bridge classes in original packages to maintain
compatibility while transitioning:

```java
@Deprecated(since = "3.6.0", forRemoval = true)
package net.lab1024.sa.common.core.domain;
public class ResponseDTO<T> extends net.lab1024.sa.foundation.domain.response.ResponseDTO<T> {
    // Delegates to parent class in new package
}
```

Bridge classes maintained:
- ResponseDTO, PageParam, PageResult (used by 50+ files)
- BaseEnum interface (extended by 30+ enums)
- ErrorCode hierarchy (used by 29+ files)
- StringConst, RequestHeaderConst (used by 17+ files)
- BusinessException (used by 6+ files)

### Files Deleted (11 duplicates eliminated)
- base/common/validator/enumeration/CheckEnum.java (duplicate)
- base/common/validator/enumeration/EnumValidator.java (duplicate)
- base/common/controller/SupportBaseController.java (duplicate)
- base/core/util/SmartIpUtil.java (migrated to ip-geolocation module)
- web/json/serializer/LongJsonSerializer.java (duplicate)
- web/json/serializer/BigDecimalNullZeroSerializer.java (duplicate)
- web/json/serializer/enumeration/EnumSerializer.java (duplicate)
- web/json/serializer/enumeration/EnumSerialize.java (duplicate)
- web/json/deserializer/LongJsonDeserializer.java (duplicate)
- common/core/enumeration/DataTypeEnum.java (enum bridge pattern failed)
- common/core/util/SmartResponseUtil.java (migrated to web module)

### Import Path Migrations (50+ files updated)
Systematic updates from old to new package structures:
- CheckEnum: base.common.validator → foundation.validation.annotation
- SwaggerTagConst: base.constant → base.swagger.constant
- SmartRequestUtil/SmartResponseUtil: base.core.util → base.web.util
- IpGeolocationUtil: base.core.util.SmartIpUtil → foundation.ipgeo.util
- PageResult, ResponseDTO: common.core.domain → foundation.domain.*

### foundation/core Module Cleanup
Reduced scope to legitimate core utilities:
- SmartBeanUtil (permanent - used by 30+ files)
- CoreAutoConfiguration (permanent - component scanning)
- SmartEnumUtil, SmartBigDecimalUtil, SmartLocalDateUtil, SmartStringUtil
- NoNeedLogin annotation (Sa-Token integration)
- ErrorCodeRangeContainer, ErrorCodeRegister
- SystemEnvironmentConfig, YamlProcessor
- DataScopePlugin, UserPermission, SystemEnvironment

Removed heavy dependencies:
- Spring Web (moved to infrastructure/web)
- MyBatis Plus (moved to infrastructure/mybatis)
- ip2region (moved to foundation/ip-geolocation)
- Apache POI, FastExcel (future: foundation/excel module)

## Documentation Changes

### Kafka Documentation Reorganization
- Moved 59 Kafka documentation files to docs/bak/kafka/
- Preserves historical documentation while cleaning up main docs structure
- Files moved:
  - Advanced guides (custom-listeners, extending-framework, idempotency,
    schema-registry, transactions)
  - Architecture docs (batch-processing, dead-letter-queue, message-flow,
    module-structure, overview)
  - Examples (aggregator, basic, batch, dlq, docker-compose)
  - Getting started (hello-world, quick-reference, quick-start)
  - Guides (batch-operations, best-practices, configuration, consumer,
    error-handling, producer)
  - Operations (backup-recovery, deployment, health-checks, monitoring,
    performance-tuning)
  - Reference (api-reference, configuration-reference, error-codes,
    migration-guide)
  - Testing (integration-testing, testing-strategy, unit-testing,
    verification-framework)
  - Troubleshooting (common-issues, debugging-tips, diagnostic-guide, faq)

### Updated Documentation
- .agent/rules/12-pmd-rules.md - PMD rule clarifications
- .agent/rules/13-spotbugs-rules.md - SpotBugs exclusion patterns
- .claude/agents/java-architect.md - Updated agent guidelines
- CLAUDE.md - Updated quick reference with new module structure
- docker/ - Configuration examples updated

### Testing Infrastructure Added
- sa-admin/src/test/java/net/lab1024/sa/admin/BaseControllerTest.java
- sa-admin/src/test/java/net/lab1024/sa/admin/BaseIntegrationTest.java
- sa-admin/src/test/java/net/lab1024/sa/admin/BaseUnitTest.java
- docs/testing/integration/ - Integration testing guides
- docs/testing/examples/ - Test examples

## Build Configuration Updates

- settings.gradle.kts: Added 4 new foundation modules
- Root build.gradle.kts: Quality checks configuration
- 28+ module build.gradle.kts: Dependency updates

## Verification Results

✅ All modules compile successfully
✅ No circular dependencies detected
✅ Backward compatibility maintained through bridge classes
✅ Full build succeeded: sa-base + sa-admin
✅ 306 files staged, 1,082 insertions(+), 31,645 deletions(-)

## Migration Path (Future Phases)

Three-phase deletion strategy documented in plan file:
1. Phase 1 (8-12 hours): Migrate all imports to new paths
2. Phase 2 (2-3 hours): Add deletion warnings to bridge classes
3. Phase 3 (1-2 hours): Delete bridge classes in v4.0.0

Current recommendation: **Keep all bridge classes** for smooth transition.

## References

- Plan file: C:\Users\ron.chang\.claude\plans\curried-honking-abelson.md
- Architecture test: sa-admin/src/test/java/net/lab1024/sa/admin/ArchitectureTest.java
- Build commands: ./gradlew clean build
- Test commands: ./gradlew :sa-admin:test --tests ArchitectureTest

+++ smart-admin-api-java21-springboot3/sa-base/foundation/api-encrypt/src/main/java/net/lab1024/sa/common/apiencrypt/advice/SmartEncryptResponseAdvice.java
@@ -3,7 +3,7 @@ package net.lab1024.sa.common.apiencrypt.advice;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.extern.slf4j.Slf4j;
 import net.lab1024.sa.common.core.domain.ResponseDTO;
-import net.lab1024.sa.common.core.enumeration.DataTypeEnum;
+import net.lab1024.sa.foundation.domain.enumeration.DataTypeEnum;
 import org.springframework.stereotype.Component;
 
 /**
@@ -32,7 +32,7 @@ public class SmartEncryptResponseAdvice extends EncryptResponseAdvice {
                 .encrypt(getObjectMapper().writeValueAsString(responseDTO.getData()));
         ResponseDTO<Object> result = (ResponseDTO<Object>) responseDTO;
         result.setData(encrypt);
-        result.setDataType(DataTypeEnum.ENCRYPT.getValue());
+        result.setDataType((Integer) DataTypeEnum.ENCRYPT.getValue());
         return result;
       } catch (JsonProcessingException e) {
         log.error("Failed to encrypt response", e);
